MODIFY Scripts Float to Int

+++ TP_TD_Train/TD_Exo_Train/Assets/TD/Scripts/DamageableDetector.cs
@@ -74,7 +74,7 @@
 			}
 		}
 
-		private void Damageable_OnDamageTaken(Damageable caller, float currentHealth, float damageTaken)
+		private void Damageable_OnDamageTaken(Damageable caller, int currentHealth, int damageTaken)
 		{
 			if (currentHealth <= 0)
 			{
